requirements.txt
# Core dependencies
openai-agents>=0.6.2
python-dotenv>=1.1.1
streamlit>=1.50.0

# Web scraping
requests==2.32.5
feedparser==6.0.12
beautifulsoup4==4.14.2
markdown

# Google integrations
gspread==6.2.1
google-auth==2.41.1
google-ads==28.0.0
pytrends==4.9.2
